+++ christmas_app/__init__.py
@@ -133,8 +133,8 @@ class About():
         return str(self.version)
 
 
-systemInfoObject = About(version=2.23, status='Public Release',
-                         build=20230512, version_note='comprehensive improvements')
+systemInfoObject = About(version=2.232, status='Public Release',
+                         build=20230515, version_note='fixed some typos')
 systemInfo = systemInfoObject.__str__()
 systemVersion = systemInfoObject.getSystemVersion()